+++ shared/src/commonMain/kotlin/com/vickbt/shared/ui/screens/home/HomeViewModel.kt
@@ -20,12 +20,6 @@ class HomeViewModel(private val moviesRepository: MoviesRepository) : ViewModel(
     private val _homeUiState = MutableStateFlow(HomeUiState(isLoading = true))
     val homeUiState = _homeUiState.asStateFlow()
 
-    private val _searchUiState = MutableStateFlow(SearchUiState(isLoading = false))
-    val searchUiState = _searchUiState.asStateFlow()
-
-    private val _searchQuery = MutableStateFlow("")
-    val searchQuery = _searchQuery.asStateFlow()
-
     private val coroutineExceptionHandler = CoroutineExceptionHandler { _, exception ->
         _homeUiState.update { it.copy(isLoading = false, error = exception.message) }
     }
@@ -83,21 +77,5 @@ class HomeViewModel(private val moviesRepository: MoviesRepository) : ViewModel(
                 _homeUiState.update { it.copy(error = error.message) }
             }
         }
-    }
-
-    fun searchMovie(movieName: String) = viewModelScope.launch(coroutineExceptionHandler) {
-        moviesRepository.searchMovie(movieName = movieName).collectLatest { moviesResult ->
-            moviesResult.isLoading { isLoading ->
-                _searchUiState.update { it.copy(isLoading = isLoading) }
-            }.onSuccess { movies ->
-                _searchUiState.update { it.copy(movieResults = movies) }
-            }.onFailure { error ->
-                _searchUiState.update { it.copy(error = error.message) }
-            }
-        }
-    }
-
-    fun updateSearchQuery(searchQuery: String) {
-        _searchQuery.value = searchQuery
     }
 }
